backend/requirements.txt
aiohttp==3.8.4
aiopg==1.4.0
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.95.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
kombu==5.2.4
multidict==6.0.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pydantic==1.10.7
pytz==2023.2
redis==4.5.3
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
